finished parent and children for sqlinteraction

+++ src/main/java/dev/ln13/cse360project/backend/Child.java
@@ -6,11 +6,11 @@ import java.util.Arrays;
 public class Child extends Patient {
 	private ArrayList<Parent> parents = new ArrayList<>();
 
-	public Child(String name, String dob, long heightCm, long weightKg, int restingHeartRate, long bloodPressurekPa, boolean childAccount, String pharmacyName, String perscribedMedication, String patientHistory, String visitSummary) {
+	public Child(String name, String dob, double heightCm, double weightKg, int restingHeartRate, double bloodPressurekPa, boolean childAccount, String pharmacyName, String perscribedMedication, String patientHistory, String visitSummary) {
 		super(name, dob, heightCm, weightKg, restingHeartRate, bloodPressurekPa, true, pharmacyName, perscribedMedication, patientHistory, visitSummary);
 	}
 
-	public Child(int id, String name, String dob, long heightCm, long weightKg, int restingHeartRate, long bloodPressurekPa, boolean childAccount, String pharmacyName, String perscribedMedication, String patientHistory, String visitSummary) {
+	public Child(int id, String name, String dob, double heightCm, double weightKg, int restingHeartRate, double bloodPressurekPa, boolean childAccount, String pharmacyName, String perscribedMedication, String patientHistory, String visitSummary) {
 		super(id, name, dob, heightCm, weightKg, restingHeartRate, bloodPressurekPa, true, pharmacyName, perscribedMedication, patientHistory, visitSummary);
 	}
